+++ libs/reconciliation/src/lib/reconciliation.module.ts
@@ -3,6 +3,7 @@ import { MovementReconcileService } from './movement-reconcile.service';
 import { ReconciliationFindingsService } from './reconciliation-findings.service';
 import { ReconciliationQueryService } from './reconciliation-query.service';
 import { ReconciliationScannerService } from './reconciliation-scanner.service';
+import { BlindCountService } from './blind-count.service';
 import { ReconciliationController } from './reconciliation.controller';
 
 /**
@@ -11,7 +12,7 @@ import { ReconciliationController } from './reconciliation.controller';
  */
 @Module({
   controllers: [ReconciliationController],
-  providers: [MovementReconcileService, ReconciliationFindingsService, ReconciliationQueryService, ReconciliationScannerService],
+  providers: [MovementReconcileService, ReconciliationFindingsService, ReconciliationQueryService, ReconciliationScannerService, BlindCountService],
   exports: [MovementReconcileService, ReconciliationFindingsService],
 })
 export class ReconciliationModule {}